refactor(archive-preview): extract format common and comic modules
